feat(dropdown): Add listSelectedTileColor and improve listTileColor
Adds a new `listSelectedTileColor` option to the `DropDown` widget to
allow customizing the background color of the selected list tile. This provides more flexibility in styling the dropdown list.

The default value for `listTileColor` has also been changed to `null` to allow the use of transparent background for list tiles.