feat(dropdown): add border option to DropDownStyle
This change adds a new `border` property to the `DropDownStyle` class, which allows
the user to customize the border shape of the bottom sheet. The default value is
a rounded rectangle border with a top radius of 24.0.

Additionally, the `getStyle` method has been added to the `DropDown` class to
simplify the logic for retrieving the final `DropDownStyle` to be used.

The `shapeBorder` property has been removed from the `DropDown` constructor, as
the border can now be specified through the `DropDownStyle`.