feat(drop_down): add new constructors for DropDown
The changes in this commit introduce new constructors for the `DropDown` widget, making it more flexible and easier to use. The new constructors are:

- `DropDown.items`: Allows creating a `DropDown` using a list of `SelectedListItem`s.
- `DropDown.raw`: Allows creating a `DropDown` using a list of items, which will be converted to `SelectedListItem`s.
- `DropDown.future`: Allows creating a `DropDown` using a future that will return a list of `SelectedListItem`s.
- `DropDown.rawFuture`: Allows creating a `DropDown` using a future that will return a list of items, which will be converted to `SelectedListItem`s.

These new constructors provide a more intuitive and flexible way to create `DropDown` widgets, making it easier for developers to integrate the widget into their applications.