feat(drop_down): add option to submit on max selection
This change adds a new option `submitOnMaxSelectionReached` to the `DropDownOptions` class. When this option is set to `true`, the drop down list will automatically submit the selection when the maximum number of selected items is reached.

This feature is useful for scenarios where the user should not be able to select more items than the allowed limit, and the form should be submitted as soon as the limit is reached.

The change also updates the `_DropDownBodyState` class to check if `submitOnMaxSelectionReached` is `true` and call `_submitMultiple(list.selected)` when the maximum selection limit is reached.